requirements.txt
tensorflow==2.2.0
Pillow==7.1.2